+++ bin/generate_from_arb.dart
@@ -3,6 +3,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart=2.10
+
 /// A main program that takes as input a source Dart file and a number
 /// of ARB files representing translations of messages from the corresponding
 /// Dart file. See extract_to_arb.dart and make_hardcoded_translation.dart.