+++ lib/src/icu_parser.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart=2.10
+
 // ignore_for_file: avoid_dynamic_calls
 
 /// Contains a parser for ICU format plural/gender/select format for localized