+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Allow null safe code when parsing.
   * Update analyzer dependency.
   * Upgrade to `package:lints/recommended.yaml`.
+  * Initial null safety conversion.
 
 ## 0.17.10+1
   * Generate code that passes analysis with `implicit-casts: false`.